Move getThankYouUrl memoizing to custom hook
